+++ changelog.md
@@ -6,6 +6,11 @@ All notable changes to this project will be documented here.
 
 Nothing yet.
 
+## [0.9.4] - April 4th, 2024
+- Fix bug where vanish mode was disabled during brush/erase mode
+- Align `undo` behavior for polygons and polylines with that of other annotations (once complete, the undo will delete the entire annotation rather than undoing only the last point)
+- `Escape` keybind to cancel an in-progress annotation
+
 ## [0.9.3] - March 22nd, 2024
 - Fix bug where erase mode could perist when changing annotation mode
 - When user provides an `initial_line_size`, use that as the default size for new annotations, independent of zoom